Reorganize Unitary TestSet
